Refactored drawer item list
