+++ Mobile/lib/Components/Navigation.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mobile/Components/TaskList.dart';
 import 'package:mobile/Components/TaskSelectedList.dart';
+import 'package:mobile/NavigationComponents/DrawerItem.dart';
 import 'package:mobile/NavigationComponents/NavigationDrawer.dart';
 import 'package:mobile/config/app_pages.dart';
 import 'package:mobile/pages/Register.dart';
@@ -19,68 +20,93 @@ class NavigationComponent extends StatefulWidget {
 
 class _NavigationComponentState extends State<NavigationComponent> {
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
-  RootAppState? appState;
+  late RootAppState _appState;
 
-  Map<AppPages, Title> _getTitles() {
-    return {
-      AppPages.home: Title('Home', Icons.home, Home()),
-      AppPages.userProfile: Title('Profile', Icons.person, UserProfilePage()),
-      AppPages.updateUserProfile:
-          Title('Edit Profile', Icons.settings, UpdateUserProfilePage(), false),
-      AppPages.leaderboard:
-          Title('Leaderboard', Icons.leaderboard, LeaderboardPage()),
-      AppPages.AssignedTasks: Title('Assigned Tasks', Icons.list,
-          TaskSelectedList(type: TasklistType.Assigned)),
-      AppPages.AvailableTasks: Title('Available Tasks', Icons.list,
-          TaskSelectedList(type: TasklistType.Available)),
-      AppPages.CompletedTasks: Title('Completed Tasks', Icons.list,
-          TaskSelectedList(type: TasklistType.Completed)),
-      AppPages.PendingTasks: Title('Pending Tasks', Icons.list,
-          TaskSelectedList(type: TasklistType.Pending)),
-      AppPages.none: Title('Logout', Icons.logout, Login(), true, _logout),
-      AppPages.login: Title('Logout', Icons.logout, Login(), false),
-      AppPages.register: Title('Logout', Icons.logout, Register(), false),
-    };
-  }
+  Map<AppPages, DrawerItem> _titles = {};
 
   void _logout() {
-    appState?.logout();
-    _onItemTapped(AppPages.login);
+    _appState.logout();
+    _appState.switchPage(AppPages.login);
   }
 
-  Color? _getSelectedColor(AppPages index) {
-    return appState?.page == index ? Colors.blue : null;
-  }
+  @override
+  void initState() {
+    super.initState();
 
-  void _onItemTapped(AppPages appPage) {
-    setState(() {
-      appState?.switchPage(appPage);
-    });
-    // Close the drawer after selecting an item
-    Navigator.pop(context);
+    _titles = {
+      AppPages.home: DrawerItem(
+        title: 'Home',
+        icon: Icons.home,
+        page: Home(),
+      ),
+      AppPages.userProfile: DrawerItem(
+        title: 'Profile',
+        icon: Icons.person,
+        page: UserProfilePage(),
+      ),
+      AppPages.leaderboard: DrawerItem(
+        title: 'Leaderboard',
+        icon: Icons.leaderboard,
+        page: LeaderboardPage(),
+      ),
+      AppPages.assignedTasks: DrawerItem(
+        title: 'Assigned Tasks',
+        icon: Icons.list,
+        page: TaskSelectedList(type: TasklistType.Assigned),
+      ),
+      AppPages.availableTasks: DrawerItem(
+        title: 'Available Tasks',
+        icon: Icons.list,
+        page: TaskSelectedList(type: TasklistType.Available),
+      ),
+      AppPages.completedTasks: DrawerItem(
+        title: 'Completed Tasks',
+        icon: Icons.list,
+        page: TaskSelectedList(type: TasklistType.Completed),
+      ),
+      AppPages.pendingTasks: DrawerItem(
+        title: 'Pending Tasks',
+        icon: Icons.list,
+        page: TaskSelectedList(type: TasklistType.Pending),
+      ),
+      AppPages.none: DrawerItem(
+        title: 'Logout',
+        icon: Icons.logout,
+        action: _logout,
+      ),
+      AppPages.login: DrawerItem(
+        title: 'Login',
+        icon: Icons.logout,
+        page: Login(),
+        show: false,
+      ),
+      AppPages.register: DrawerItem(
+        title: 'Register',
+        icon: Icons.logout,
+        page: Register(),
+        show: false,
+      ),
+    };
   }
 
   @override
   Widget build(BuildContext context) {
-    if (appState == null) {
-      appState = context.watch<RootAppState>();
-    }
-    Map<AppPages, Title> titles = _getTitles();
-    bool? isLoggedIn = appState?.isLoggedInSync();
+    _appState = context.watch<RootAppState>();
+    bool isLoggedIn = _appState.isLoggedInSync();
 
-    if (!(isLoggedIn ?? false)) {
+    if (!isLoggedIn) {
       return Scaffold(
-        body: Center(child: titles[appState?.page]?.page),
+        body: Center(child: _titles[_appState.page]?.page),
       );
     }
 
     return Scaffold(
       key: _scaffoldKey,
       // if not logged in or root app state is null don't add drawer.
-      appBar: !(isLoggedIn ?? false)
+      appBar: !isLoggedIn
           ? null
           : AppBar(
-              title: Text('Title'),
+              title: Text(_titles[_appState.page]?.title ?? ''),
               backgroundColor: Color(0xFFF5C53A),
               actions: <Widget>[
                 // maybe use this for user profile -_-
@@ -92,20 +118,12 @@ class _NavigationComponentState extends State<NavigationComponent> {
                 // ),
               ],
             ),
-      drawer: NavDrawer(),
+      drawer: NavDrawer(
+        titles: _titles,
+      ),
       body: Center(
-        child: titles[appState?.page]?.page,
+        child: _titles[_appState.page]?.page,
       ),
     );
   }
-}
-
-class Title {
-  String title;
-  IconData icon;
-  bool show = true;
-  StatefulWidget page;
-  VoidCallback? action;
-
-  Title(this.title, this.icon, this.page, [this.show = true, this.action]);
 }
